refactor: organize packages and renamed widgets
