+++ lib/data/network/source/local/movies/local_data_source.dart
@@ -1,0 +1,13 @@
+import '../../../../../domain/entities/movie.dart';
+
+abstract class LocalDataSource {
+
+  Future<List<Movie>> getFavorites();
+
+  Future<void> addToFavorites(Movie movie);
+
+  Future<bool> isFavorite(int id);
+
+  Future<void> removeFromFavorites(Movie movie);
+}
+
