+++ lib/data/network/source/remote/movie_data_source.dart
@@ -1,4 +1,4 @@
-import '../../domain/entities/movie.dart';
+import '../../../../domain/entities/movie.dart';
 
 abstract class MovieDataSource {
   Future<List<Movie>> getPopularMovies();
