+++ lib/data/service/movie_service_api.dart
@@ -1,6 +1,6 @@
 import 'package:dio/dio.dart';
 import 'package:retrofit/http.dart';
-import '../models/movies_response.dart';
+import '../models/movies/movies_response.dart';
 part 'movie_service_api.g.dart';
 
 @RestApi(baseUrl: 'https://api.themoviedb.org/3')
