+++ lib/domain/usecase/movies/add_to_favorites_use_case.dart
@@ -1,5 +1,5 @@
-import '../entities/movie.dart';
-import '../repository/movie_repository.dart';
+import '../../entities/movie.dart';
+import '../../repository/movie_repository.dart';
 
 class AddToFavoritesUseCase {
   final MoviesRepository moviesRepository;
