+++ lib/domain/usecase/movies/get_popular_movie_use_case.dart
@@ -1,5 +1,6 @@
-import '../entities/movie.dart';
-import '../repository/movie_repository.dart';
+
+import '../../entities/movie.dart';
+import '../../repository/movie_repository.dart';
 
 class GetPopularMoviesUseCase {
   final MoviesRepository repository;
