+++ lib/domain/usecase/movies/is_favorite_use_case.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
-import '../entities/movie.dart';
-import '../repository/movie_repository.dart';
+import '../../entities/movie.dart';
+import '../../repository/movie_repository.dart';
 
 class IsFavoriteUseCase {
   final MoviesRepository moviesRepository;
