+++ lib/domain/usecase/movies/remove_from_favorite_use_case.dart
@@ -1,6 +1,5 @@
-
-import '../entities/movie.dart';
-import '../repository/movie_repository.dart';
+import '../../entities/movie.dart';
+import '../../repository/movie_repository.dart';
 
 class RemoveFromFavoriteUseCase {
   final MoviesRepository moviesRepository;
